feat(crypto): add timingSafeEqual constant-time compare helper (#594)
Introduce timing-safe byte comparison for secrets and document when to use it across security references.
