+++ packages/crypto/src/index.ts
@@ -55,4 +55,5 @@ export { deriveKeypairFromMnemonic } from './key-derivation';
 export { signTransaction, verifySignature } from './signing';
 
 // Constant-time comparison
+export { timingSafeEqual } from './timing-safe';
 export { constantTimeEqual } from './compare';
